html/semantics/forms/the-input-element/mobiles.html

<!DOCTYPE html>
<html>

  <head>
    <title>Mobile Inputs</title>
     <meta name=viewport content="width=device-width, maximum-scale=1.0, userscalable=no" />
     <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
 	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	  <link rel="help" href="http://www.w3.org/TR/html5/the-input-element.html#dom-input-type">
   <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>

  		<h1>Mobile Inputs</h1>
  		<div style="display: none">
	  		<input type="text" />
	  		<input type="number" />
	  		<input type="email" />
	  		<input type="tel" />
	  		<input type="url" />
		</div>
		
    <div id="log">
    </div>
	
	<script type="text/javascript">
		
		test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "text")}, "text type support on input element");
		test(function() {assert_equals(document.getElementsByTagName("input")[1].type, "number")}, "number type support on input element");
		test(function() {assert_equals(document.getElementsByTagName("input")[2].type, "email")}, "email type support on input element");  
		test(function() {assert_equals(document.getElementsByTagName("input")[3].type, "tel")}, "tel type support on input element");
		test(function() {assert_equals(document.getElementsByTagName("input")[4].type, "url")}, "url type support on input element");

	</script>
	
  </body>
  
</html>
